Style changes and general fixes for ShareModal in Cordova iOS Moved a lot of copied code from ShareModal and ShareButtonFooter to common files Fixed the red error on Ballot complaining that lists must have unique keys
