+++ src/js/components/Share/shareButtonCommon.jsx
@@ -1,5 +1,117 @@
-import { cordovaOpenSafariView } from '../../common/utils/cordovaUtils';
+import { FileCopyOutlined } from '@mui/icons-material';
+import { Button } from '@mui/material';
+import React, { Suspense } from 'react';
+import { FacebookIcon, FacebookShareButton, TwitterIcon, TwitterShareButton } from 'react-share'; // EmailIcon, EmailShareButton,
+import styled from 'styled-components';
+import OpenExternalWebSite from '../../common/components/Widgets/OpenExternalWebSite';
+import { cordovaOpenSafariView, hasDynamicIsland, hasIPhoneNotch, isAndroid } from '../../common/utils/cordovaUtils';
+import { isCordova } from '../../common/utils/isCordovaOrWebApp';
 import { openSnackbar } from '../Widgets/SnackNotifier';
+import ShareModalOption from './ShareModalOption';
+
+export const shareStyles = () => ({
+  dialogPaper: {
+    // marginTop: hasIPhoneNotch() ? 68 : 48,
+    '@media (min-width: 576px)': {
+      maxWidth: '600px',
+      width: '90%',
+      height: 'fit-content',
+      margin: '0 auto',
+      minWidth: 0,
+      minHeight: 0,
+      transitionDuration: '.25s',
+    },
+    minWidth: '100%',
+    maxWidth: '100%',
+    width: '100%',
+    minHeight: '100%',
+    maxHeight: '100%',
+    height: '100%',
+    margin: '0 auto',
+  },
+  dialogContent: {
+    padding: '0 24px 36px 24px',
+    background: 'white',
+    height: 'fit-content',
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+    justifyContent: 'center',
+    '@media(max-width: 576px)': {
+      justifyContent: 'flex-start !important',
+    },
+  },
+  buttonDefault: {
+    padding: '0 12px',
+    width: '100%',
+    boxShadow: 'none !important',
+    borderRadius: '0 !important',
+    height: '45px !important',
+  },
+  buttonDefaultCordova: {
+    padding: '0 12px',
+    width: '100%',
+    boxShadow: 'none !important',
+    borderRadius: '0 !important',
+    height: '35px !important',
+  },
+  backButton: {
+    marginBottom: 6,
+    marginLeft: -8,
+    // paddingTop: 0,
+    // paddingBottom: 0,
+  },
+  backButtonIcon: {
+    fontSize: 14,
+    fontWeight: 'bold',
+  },
+  cancelButton: {
+    marginTop: 12,
+  },
+  closeButton: {
+    marginLeft: 'auto',
+  },
+  closeButtonAbsolute: {
+    position: 'absolute',
+    right: 5,
+    top: hasIPhoneNotch() || hasDynamicIsland() ? 30 : 3,
+  },
+  informationIcon: {
+    color: '#999',
+    width: 20,
+    height: 20,
+    marginTop: '-3px',
+    marginRight: 3,
+  },
+  informationIconInButton: {
+    color: '#999',
+    width: 16,
+    height: 16,
+    marginTop: '-3px',
+    marginLeft: 3,
+  },
+  menuItem: {
+    zIndex: '9 !important',
+    padding: '0 !important',
+    marginBottom: '-2px !important',
+    paddingBottom: '1px !important',
+    '&:last-child': {
+      paddingBottom: '0 !important',
+      paddingTop: '1px !important',
+    },
+    '&:hover': {
+      background: '#efefef',
+    },
+  },
+  previewButton: {
+    marginTop: 0,
+  },
+  shareIcon: {
+    transform: 'scaleX(-1)',
+    position: 'relative',
+    top: -1,
+  },
+});
 
 export function androidFacebookClickHandler (linkToBeShared) {
   // react-share in Cordova for Android, navigates to the URL instead of opening a "tab" with a return 'X' button
@@ -50,5 +162,153 @@ export function cordovaSocialSharingByEmail (subject, linkToBeShared, handleClos
     null,               // FILES: can be null, a string, or an array
     onEmailSendSuccess, // called when sharing worked, but also when the user cancelled sharing via email. On iOS, the callbacks' boolean result parameter is true when sharing worked, false if cancelled. On Android, this parameter is always true so it can't be used). See section "Notes about the successCallback" below.
     onEmailSendError,   // called when sh*t hits the fan
+  );
+}
+
+const TextH3 = styled('h3')`
+  font-weight: normal;
+  font-size: 16px;
+  color: black !important;
+  padding: 6px;
+`;
+
+const ShareWrapper = styled('div')`
+  cursor: pointer;
+  display: block !important;
+  // margin-bottom: 12px;
+  @media (min-width: 600px) {
+    flex: 1 1 0;
+  }
+  height: 100%;
+  text-align: center;
+  text-decoration: none !important;
+  color: black !important;
+  transition-duration: .25s;
+  &:hover {
+    text-decoration: none !important;
+    color: black !important;
+    transform: scale(1.05);
+    transition-duration: .25s;
+  }
+  @media (max-width: 600px) {
+    width: 33.333%;
+  }
+`;
+
+/* eslint-disable react/prop-types */
+// React functional component example
+export function ShareWeVoteFriends (props) {
+  const { onClickFunction } = props;
+  return (
+    <ShareModalOption
+      backgroundColor="#0834cd"
+      icon={<img src="../../../img/global/svg-icons/we-vote-icon-square-color.svg" alt="" />}
+      // urlToShare={linkToBeShared}
+      noLink
+      onClickFunction={onClickFunction}
+      title="We Vote friends"
+      uniqueExternalId="shareButtonFooter-Friends"
+    />
+  );
+}
+
+export function SharePreviewFriends (props) {
+  const { classes, linkToBeShared } = props;
+  return (
+    <Suspense fallback={<></>}>
+      <OpenExternalWebSite
+        linkIdAttribute="allOpinions"
+        url={linkToBeShared}
+        target="_blank"
+        // title={this.props.title}
+        className="u-no-underline"
+        body={(
+          <Button className={classes.previewButton} variant="outlined" fullWidth color="primary">
+            Preview what your friends will see
+          </Button>
+        )}
+        style={isCordova() ? { display: 'none' } : {}}
+      />
+    </Suspense>
+  );
+}
+
+// React functional component example
+export function ShareFacebook (props) {
+  const { titleText, saveActionShareButtonFacebook, linkToBeShared } = props;
+  return (
+    <ShareWrapper>
+      <div id="androidFacebook"
+           onClick={() => isAndroid() && androidFacebookClickHandler(`${linkToBeShared}&t=WeVote`)}
+      >
+        <FacebookShareButton
+          className="no-decoration"
+          id="shareFooterFacebookButton"
+          onClick={saveActionShareButtonFacebook}
+          quote={titleText}
+          url={`${linkToBeShared}&t=WeVote`}
+          windowWidth={750}
+          windowHeight={600}
+          disabled={isAndroid()}
+          disabledStyle={isAndroid() ? { opacity: 1 } : {}}
+        >
+          <FacebookIcon
+            bgStyle={{ background: '#3b5998' }}
+            round="True"
+            size={68}
+          />
+          <TextH3>
+            Facebook
+          </TextH3>
+        </FacebookShareButton>
+      </div>
+    </ShareWrapper>
+  );
+}
+
+// React functional component example
+export function ShareTwitterAndCopy (props) {
+  const { titleText, saveActionShareButtonTwitter, saveActionShareButtonCopy, linkToBeSharedTwitter, linkToBeSharedCopy } = props;
+  return (
+    <>
+      <ShareWrapper>
+        <div id="androidTwitter"
+             onClick={() => isAndroid() &&
+               androidTwitterClickHandler(linkToBeSharedTwitter)}
+        >
+          <TwitterShareButton
+            className="no-decoration"
+            id="shareFooterTwitterButton"
+            onClick={saveActionShareButtonTwitter}
+            title={titleText}
+            url={`${linkToBeSharedTwitter}`}
+            windowWidth={750}
+            windowHeight={600}
+            disabled={isAndroid()}
+            disabledStyle={isAndroid() ? { opacity: 1 } : {}}
+          >
+            <TwitterIcon
+              bgStyle={{ background: '#38A1F3' }}
+              round="True"
+              size={68}
+            />
+            <TextH3>
+              Twitter
+            </TextH3>
+          </TwitterShareButton>
+        </div>
+      </ShareWrapper>
+      <ShareWrapper>
+        <ShareModalOption
+          backgroundColor="#2E3C5D"
+          copyLink
+          icon={<FileCopyOutlined />}
+          urlToShare={linkToBeSharedCopy}
+          onClickFunction={saveActionShareButtonCopy}
+          title="Copy link"
+          uniqueExternalId="shareButtonFooter-CopyLink"
+        />
+      </ShareWrapper>
+    </>
   );
 }
